ComorbidityIndexes.jl

This package implements several comorbidity indexes in ICD-10 coding.

Currently available are the Charlson and Elixhauser comorbidities, using the Quan (2005) ICD-10 codes, and the Multipurpose Australian Comorbidity Scoring System (MACSS) using ICD-10-AM codes.

Summary scores for the Charlson comorbidity index can be calculated using the original Charlson weights or the updated Quan (2011) weights. A summary score for the Elixhauser comorbidities can be calculated using the van Walraven (2009) weights.

ComorbidityIndexes.flagcomorbidities — Method

flagcomorbidities(icdcodes, comorbidities)

Flag comorbidities in an array of ICD-10 codes.

comorbidities can be :elixhauser or :charlson for Quan ICD-10 versions of Elixhauser and Charlson comorbidities, respectively, or :macss for the Multipurpose Australian Comorbidity Scoring System.

Returns Dict giving true or false for each comorbidity.

ComorbidityIndexes.scorecomorbidity — Method

scorecomorbidity(comorbidities, score)

Compute summary comorbidity score.

comorbidities should be the output from flagcomorbidities(@ref). score should be :charlson for original Charlson weights, :charlson_quan2011 for Quan's updated Charlson weights, or :vanwalraven for the van Walraven weights for the Elixhauser comorbidities.

It is your responsibility to request a score compatible with the comorbidity flags you provide.